Move Devices to compose service module
