+++ modules/services/compose/src/main/java/au/com/shiftyjelly/pocketcasts/compose/Devices.kt
@@ -1,6 +1,6 @@
-package au.com.shiftyjelly.pocketcasts.sharing.ui
+package au.com.shiftyjelly.pocketcasts.compose
 
-internal object Devices {
+object Devices {
     const val PortraitRegular = "spec:width=800px,height=1600px,dpi=320"
     const val PortraitSmall = "spec:width=720px,height=1280px,dpi=320"
     const val PortraitTablet = "spec:width=1600px,height=2560px,dpi=276"
